requirements.txt
google-generativeai>=0.5.4
google-cloud-storage
python-dotenv
Pillow
ffmpeg-python
gtts